Add a timeout for kindnetd when listing nodes
Here timeout is 1 second to speed up control-plane node to be ready as
fast as possible. because node is waiting for container runtime to be
ready which needs cni to be ready, i.e. kindnetd to be ready.

Otherwise in first time to list nodes kindnetd will encounter network
timeout problem whose timeout is half minute.